Add comments in the TypeScript definitions
