+++ lib/VoxeetSDK.ts
@@ -1,78 +1,14 @@
-import UserInfo from "./UserInfo";
 import VoxeetMedia from "./VoxeetMedia";
-import { Configuration } from "./Configurations";
-/*
- *
- * VoxeetUXKit Cordova
- * Copyright (C) 2020
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
-
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
-
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- *
-*/
+import ConferenceUser from "./types/ConferenceUser";
+import { default as Configuration } from "./types/Configurations";
+import { CreateOptions, CreateResult } from './types/CreateConference';
+import { JoinOptions } from './types/JoinConference';
 
 /**
  * This class enable interaction with VoxeetUXKit
- * @constructor
  */
-const exec:any = require('cordova/exec');
-const SERVICE:string = 'Voxeet';
-
-export enum RTCPMode {
-    WORST = "worst",
-    BEST = "best"
-}
-
-export enum Mode {
-    STANDARD = "standard",
-    PUSH = "push"
-}
-
-export enum Codec {
-    VP8 = "VP8",
-    H264 = "H264"
-}
-
-export { Configuration as Configuration} from "./Configurations";
-export { Users as Users} from "./Configurations";
-export { Overlay as Overlay} from "./Configurations";
-export { ActionBar as ActionBar} from "./Configurations";
-
-export interface CreateParameters {
-    ttl?: number;
-    rtcpMode?: RTCPMode; //best / worst, default => worst
-    mode?: Mode; // push / standard, default => standard
-    videoCodec?: Codec; //default VP8
-    liveRecording?: boolean; //default false
-}
-
-export interface CreateOptions {
-    alias?: string;
-    params?: CreateParameters;
-}
-
-export enum UserType {
-    USER = "user",
-    LISTENER = "listener"
-}
-
-export interface JoinUserInfo {
-    type?:  UserType;
-}
-
-export interface JoinOptions {
-    user?: JoinUserInfo;
-}
+const exec: any = require('cordova/exec');
+const SERVICE: string = 'Voxeet';
 
 export interface RefreshCallback {
     (): void;
@@ -86,11 +22,10 @@ class Voxeet_ {
 
     public VoxeetMedia: VoxeetMedia;
 
-    refreshAccessTokenCallback: RefreshCallback|null = null;
-    refreshToken: TokenRefreshCallback|undefined; 
+    refreshAccessTokenCallback: RefreshCallback | null = null;
+    refreshToken: TokenRefreshCallback | undefined; 
 
     constructor() {
-
         this.VoxeetMedia = new VoxeetMedia();
         
         this.refreshAccessTokenCallback = () => {
@@ -103,13 +38,23 @@ class Voxeet_ {
         }
     }
 
-    initialize(consumerKey: string, consumerSecret: string): Promise<any> {
+    /**
+     * Initializes the SDK using the customer key and secret.
+     * @param consumerKey Consumer Key
+     * @param consumerSecret Consumer Secret
+     */
+    initialize(consumerKey: string, consumerSecret: string): Promise<string> {
         return new Promise((resolve, reject) => {
             exec(resolve, reject, SERVICE, 'initialize', [consumerKey, consumerSecret]);
         });
     }
 
-    initializeToken(accessToken: string|undefined, refreshToken: TokenRefreshCallback) {
+    /**
+     * Initializes the SDK with an access token that is provided by the customer backend communicating with Voxeet servers.
+     * @param accessToken Access token
+     * @param refreshToken Callback to get a new access token after it expires
+     */
+    initializeToken(accessToken: string | undefined, refreshToken: TokenRefreshCallback): Promise<string> {
         return new Promise((resolve, reject) => {
             this.refreshToken = refreshToken;
             exec(this.refreshAccessTokenCallback, (err: Error) => {}, SERVICE, 'refreshAccessTokenCallback', []);
@@ -117,172 +62,260 @@ class Voxeet_ {
         });
     }
 
-    connect(userInfo: UserInfo): Promise<any> {
+    /**
+     * Opens a new session.
+     * @param userInfo Participant information
+     */
+    connect(userInfo: ConferenceUser): Promise<string> {
         return new Promise((resolve, reject) => {
             exec(resolve, reject, SERVICE, 'connect', [userInfo.json()]);
         });
     }
 
-    disconnect(): Promise<any>  {
+    /**
+     * Closes the current session.
+     */
+    disconnect(): Promise<string>  {
         return new Promise((resolve, reject) => {
             exec(resolve, reject, SERVICE, 'disconnect', []);
         });
     }
 
-    create(options: CreateOptions): Promise<any> {
+    /**
+     * Creates a conference.
+     * @param options Options to use to create the conference
+     */
+    create(options: CreateOptions): Promise<CreateResult> {
         return new Promise((resolve, reject) => {
             exec(resolve, reject, SERVICE, 'create', [options]);
         });
     }
 
-    join(conferenceId: string, options: JoinOptions = {}): Promise<any>  {
+    /**
+     * Joins the conference.
+     * @param conferenceId Id of the conference to join
+     * @param options Options to use to join the conference
+     */
+    join(conferenceId: string, options: JoinOptions = {}): Promise<string>  {
         return new Promise((resolve, reject) => {
             exec(resolve, reject, SERVICE, 'join', [conferenceId, options]);
         });
     }
 
-    broadcast(conferenceId: string): Promise<any>  {
+    /**
+     * Joins the conference in the broadcaster mode which allows transmitting audio and video.
+     * @param conferenceId Id of the conference to join
+     */
+    broadcast(conferenceId: string): Promise<string>  {
         return new Promise((resolve, reject) => {
             exec(resolve, reject, SERVICE, 'broadcast', [conferenceId]);
         });
     }
 
-    leave(): Promise<any>  {
+    /**
+     * Leaves the conference.
+     */
+    leave(): Promise<string>  {
         return new Promise((resolve, reject) => {
             exec(resolve, reject, SERVICE, 'leave', []);
         });
     }
 
-    invite(conferenceId: string, participants: UserInfo[]): Promise<any> {
+    /**
+     * Invite a participant to the conference.
+     * @param conferenceId Id of the conference to invite the participant to
+     * @param participants List of participants to invite
+     */
+    invite(conferenceId: string, participants: Array<ConferenceUser>): Promise<any> {
         const array = participants ? participants.map(e => e.json()) : null;
         return new Promise((resolve, reject) => {
             exec(resolve, reject, SERVICE, 'invite', [conferenceId, array]);
         });
     }
 
+    /**
+     * Sends a broadcast message to the participants of the conference.
+     * @param message Message to send to the other participants
+     */
     sendBroadcastMessage(message: string): Promise<any> {
         return new Promise((resolve, reject) => {
             exec(null, null, SERVICE, 'sendBroadcastMessage', [message]);
-            resolve();
+            resolve(null);
         });
     }
 
+    /**
+     * Sets if you want to enable audio 3D.
+     * @param enabled True to enable audio 3D
+     */
     setAudio3DEnabled(enabled: boolean): Promise<any> {
         return new Promise((resolve, reject) => {
             exec(resolve, reject, SERVICE, 'setAudio3DEnabled', [enabled]);
         });
     }
 
+    /**
+     * Is audio 3D enabled.
+     */
     isAudio3DEnabled(): Promise<boolean> {
-        return new Promise((resolve, reject) => exec(resolve, reject, SERVICE, 'isAudio3DEnabled', []));
+        return new Promise((resolve, reject) => {
+            exec(resolve, reject, SERVICE, 'isAudio3DEnabled', []);
+        });
     }
 
+    /**
+     * Sets the UI configuration.
+     * @param configuration UI configuration
+     */
     setUIConfiguration(configuration: Configuration) {
-        return new Promise((resolve, reject) => exec(resolve, reject, SERVICE, 'setUIConfiguration', [configuration]));
+        return new Promise((resolve, reject) => {
+            exec(resolve, reject, SERVICE, 'setUIConfiguration', [configuration]);
+        });
     }
 
+    /**
+     * Sets if you want to enable the Telecom mode or not.
+     * @param enabled True to enable the Telecom mode
+     */
     setTelecomMode(enabled: boolean): Promise<any> {
         return new Promise((resolve, reject) => {
             exec(resolve, reject, SERVICE, 'setTelecomMode', [enabled]);
         });
     }
 
+    /**
+     * Is telecom mode enabled.
+     */
     isTelecomMode(): Promise<boolean> {
-        return new Promise((resolve, reject) => exec(resolve, reject, SERVICE, 'isTelecomMode', []));
+        return new Promise((resolve, reject) => {
+            exec(resolve, reject, SERVICE, 'isTelecomMode', []);
+        });
     }
 
-    appearMaximized(enabled: boolean): Promise<any> {
+    /**
+     * Sets if you want the UXKit to appear maximized or not.
+     * @param maximized True to have the UXKit to appear maximized
+     */
+    appearMaximized(enabled: boolean): Promise<void> {
         return new Promise((resolve, reject) => {
             exec(null, null, SERVICE, 'appearMaximized', [enabled]);
             resolve();
         });
     }
 
-    defaultBuiltInSpeaker(enabled: boolean): Promise<any> {
+    /**
+     * Use the built in speaker by default.
+     * @param enable True to use the built in speaker by default
+     */
+    defaultBuiltInSpeaker(enabled: boolean): Promise<void> {
         return new Promise((resolve, reject) => {
             exec(null, null, SERVICE, 'defaultBuiltInSpeaker', [enabled]);
             resolve();
         });
     }
 
-    defaultVideo(enabled: boolean): Promise<any> {
+    /**
+     * Sets the video on by default.
+     * @param enable True to turn on the video by default
+     */
+    defaultVideo(enabled: boolean): Promise<void> {
         return new Promise((resolve, reject) => {
             exec(null, null, SERVICE, 'defaultVideo', [enabled]);
             resolve();
         });
     }
 
+    /**
+     * Starts recording the conference.
+     */
     startRecording() {
         return new Promise((resolve, reject) => {
             exec(null, null, SERVICE, 'startRecording', []);
-            resolve();
+            resolve(null);
         });
     }
 
+    /**
+     * Stops recording the conference.
+     */
     stopRecording() {
         return new Promise((resolve, reject) => {
             exec(null, null, SERVICE, 'stopRecording', []);
-            resolve();
+            resolve(null);
         });
     }
 
-    /*
-     *  Android methods
+    /**
+     * Activates or disable the screen auto lock. Android only.
+     * @param activate True to activate the screen auto lock
      */
-
     screenAutoLock(enabled: boolean): Promise<any> {
         return new Promise((resolve, reject) => {
             exec(null, null, SERVICE, 'screenAutoLock', [enabled]);
-            resolve();
+            resolve(null);
         });
     }
 
+    /**
+     * @deprecated Android only.
+     */
     isUserLoggedIn(): Promise<boolean> {
-        return new Promise((resolve, reject) => exec(resolve, reject, SERVICE, 'isUserLoggedIn', []));
+        return new Promise((resolve, reject) => {
+            exec(resolve, reject, SERVICE, 'isUserLoggedIn', []);
+        });
     }
 
+    /**
+     * Checks if a conference is awaiting. Android only.
+     */
     checkForAwaitingConference(): Promise<any> {
-        return new Promise((resolve, reject) => exec(resolve, reject, SERVICE, 'checkForAwaitingConference', []));
+        return new Promise((resolve, reject) => {
+            exec(resolve, reject, SERVICE, 'checkForAwaitingConference', []);
+        });
     }
 
-    /** @deprecated */
-    startConference(conferenceId: string, participants: Array<UserInfo>): Promise<any> {
+    /** @deprecated Use join() instead. */
+    startConference(conferenceId: string, participants: Array<ConferenceUser>): Promise<any> {
         const array = participants ? participants.map(e => e.json()) : null;
         return new Promise((resolve, reject) => {
             exec(resolve, reject, SERVICE, 'startConference', [conferenceId, array]);
         });
     }
 
-    /** @deprecated */
+    /** @deprecated Use leave() instead. */
     stopConference(): Promise<any> {
-        return new Promise((resolve, reject) => exec(resolve, reject, SERVICE, 'stopConference', []));
+        return new Promise((resolve, reject) => {
+            exec(resolve, reject, SERVICE, 'stopConference', []);
+        });
     }
 
-    /** @deprecated use connect instead */
-    openSession(userInfo: UserInfo): Promise<any> {
+    /** @deprecated use connect instead. */
+    openSession(userInfo: ConferenceUser): Promise<any> {
         return new Promise((resolve, reject) => {
             exec(resolve, reject, SERVICE, 'openSession', [userInfo.json()]);
         });
     }
 
-    /** @deprecated use disconnect instead */
+    /** @deprecated use disconnect instead. */
     closeSession(): Promise<any> {
-        return new Promise((resolve, reject) => exec(resolve, reject, SERVICE, 'closeSession', []));
+        return new Promise((resolve, reject) => {
+            exec(resolve, reject, SERVICE, 'closeSession', []);
+        });
     }
 
-    //method to refresh tokens, used internally
-    onAccessTokenOk (accessToken: string) {
-      return new Promise((resolve, reject) => {
-        exec(resolve, reject, SERVICE, 'onAccessTokenOk', [accessToken]);
-      });
+    /** Internal method used to refresh tokens. */
+    onAccessTokenOk(accessToken: string) {
+        return new Promise((resolve, reject) => {
+            exec(resolve, reject, SERVICE, 'onAccessTokenOk', [accessToken]);
+        });
     }
 
-    onAccessTokenKo (errorMessage: string) {
-      return new Promise((resolve, reject) => {
-        exec(resolve, reject, SERVICE, 'onAccessTokenKo', [errorMessage]);
-      });
+    /** Internal method. */
+    onAccessTokenKo(errorMessage: string) {
+        return new Promise((resolve, reject) => {
+            exec(resolve, reject, SERVICE, 'onAccessTokenKo', [errorMessage]);
+        });
     }
 }
 
 export const Voxeet = new Voxeet_();
-//export default new Voxeet(); // will be available through Voxeet not voxeet -> fake 'singleton'
