+++ lib/types/Configurations.ts
@@ -1,0 +1,23 @@
+export interface ActionBar {
+    displayMute?: boolean,
+    displaySpeaker?: boolean,
+    displayCamera?: boolean,
+    displayScreenShare?: boolean,
+    displayLeave?: boolean
+}
+
+export interface Overlay {
+    backgroundMaximizedColor?: number,
+    backgroundMinimizedColor?: number
+}
+
+export interface Users {
+    speakingUserColor?: number,
+    selectedUserColor?: number
+}
+
+export default interface Configuration {
+    actionBar?: ActionBar,
+    overlay?: Overlay,
+    users?: Users
+}
