+++ lib/types/JoinConference.ts
@@ -1,0 +1,12 @@
+export enum UserType {
+    USER = "user",
+    LISTENER = "listener"
+}
+
+export interface JoinUserInfo {
+    type?:  UserType;
+}
+
+export interface JoinOptions {
+    user?: JoinUserInfo;
+}
