+++ lib/types/index.tsx
@@ -1,0 +1,7 @@
+export { default as ConferenceUser } from "./ConferenceUser";
+
+export { default as Configuration } from "./Configurations";
+
+export { CreateOptions, CreateResult } from "./CreateConference";
+
+export { JoinOptions } from "./JoinConference";
