+++ www/types/CreateConference.js
@@ -1,0 +1,20 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+exports.Codec = exports.Mode = exports.RTCPMode = void 0;
+var RTCPMode;
+(function (RTCPMode) {
+    RTCPMode["WORST"] = "worst";
+    RTCPMode["AVERAGE"] = "average";
+    RTCPMode["BEST"] = "best";
+})(RTCPMode = exports.RTCPMode || (exports.RTCPMode = {}));
+var Mode;
+(function (Mode) {
+    Mode["STANDARD"] = "standard";
+    Mode["PUSH"] = "push";
+})(Mode = exports.Mode || (exports.Mode = {}));
+var Codec;
+(function (Codec) {
+    Codec["VP8"] = "VP8";
+    Codec["H264"] = "H264";
+})(Codec = exports.Codec || (exports.Codec = {}));
+//# sourceMappingURL=CreateConference.js.map
