+++ www/types/JoinConference.d.ts
@@ -1,0 +1,10 @@
+export declare enum UserType {
+    USER = "user",
+    LISTENER = "listener"
+}
+export interface JoinUserInfo {
+    type?: UserType;
+}
+export interface JoinOptions {
+    user?: JoinUserInfo;
+}
