+++ www/types/JoinConference.js
@@ -1,0 +1,9 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+exports.UserType = void 0;
+var UserType;
+(function (UserType) {
+    UserType["USER"] = "user";
+    UserType["LISTENER"] = "listener";
+})(UserType = exports.UserType || (exports.UserType = {}));
+//# sourceMappingURL=JoinConference.js.map
