+++ www/types/index.jsx
@@ -1,0 +1,9 @@
+"use strict";
+var __importDefault = (this && this.__importDefault) || function (mod) {
+    return (mod && mod.__esModule) ? mod : { "default": mod };
+};
+Object.defineProperty(exports, "__esModule", { value: true });
+exports.ConferenceUser = void 0;
+var ConferenceUser_1 = require("./ConferenceUser");
+Object.defineProperty(exports, "ConferenceUser", { enumerable: true, get: function () { return __importDefault(ConferenceUser_1).default; } });
+//# sourceMappingURL=index.jsx.map
